+++ client/src/App.jsx
@@ -27,8 +27,17 @@ function App() {
       </Routes>
       
       <ToastContainer
-        position="top-right" autoClose={5000} hideProgressBar={false} newestOnTop={false}
-        closeOnClick rtl={false} pauseOnFocusLoss draggable pauseOnHover theme="colored"
+        position="top-right"
+        autoClose={3500}
+        hideProgressBar={false}
+        newestOnTop={true}
+        closeOnClick
+        rtl={false}
+        pauseOnFocusLoss
+        draggable
+        pauseOnHover
+        theme="light"
+        limit={3}
       />
     </>
   );